fix: split app.py into a seperate file to create the map and prettified the map hover text
